pai_nai_dee_backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
email-validator
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
pytest
# pytest-asyncio removed to rely on anyio's plugin
httpx
